[gvf] Refactor GVF module framework, separate nav functions from guidance/control and fix rotorcraft firmware bugs
