+++ conf/airframes/UCM/rover_steering.xml
@@ -44,10 +44,8 @@
       <configure name="USE_MAGNETOMETER" value="0"/>
     </module-->
 
-    <module name="nav"  type="rover_base">
-      <define name="ROVER_BASE_SEND_TRAJECTORY" value="FALSE"/>
-    </module>
-    <module name="gvf"  type="module">
+    <module name="nav"  type="rover_base"/>
+    <module name="gvf"  type="classic">
       <define name="GVF_OCAML_GCS" value="FALSE"/>
     </module>
