+++ conf/modules/gvf_classic.xml
@@ -1,6 +1,6 @@
 <!DOCTYPE module SYSTEM "module.dtd">
 
-<module name="gvf_module" dir="guidance/gvf">
+<module name="gvf_classic" dir="guidance/gvf">
   <doc>
     <description>Guidance algorithm for tracking smooth trajectories. The algorithm is based on the idea of stearing the vehicle to a vector field that smoothly converges to the desired trajectory.
 For more details we refer to https://wiki.paparazziuav.org/wiki/Module/guidance_vector_field .
@@ -37,21 +37,21 @@ For more details we refer to https://wiki.paparazziuav.org/wiki/Module/guidance_
           <dl_setting MAX="20" MIN="0.0" STEP="0.2" VAR="gvf_control.speed" shortname = "(ROTORCRAFT) speed"/>
         </dl_settings>
         <dl_settings NAME="Ellipse">
-          <dl_setting MAX="5" MIN="0.0" STEP="0.01" VAR="gvf_ellipse_par.ke" shortname="ell_ke" param="GVF_ELLIPSE_KE"/>
+          <dl_setting MAX="50" MIN="0.0" STEP="0.01" VAR="gvf_ellipse_par.ke" shortname="ell_ke" param="GVF_ELLIPSE_KE"/>
           <dl_setting MAX="5" MIN="0.0" STEP="0.01" VAR="gvf_ellipse_par.kn" shortname="ell_kn" param="GVF_ELLIPSE_KN"/>
           <dl_setting MAX="150" MIN="0.0" STEP="5" VAR="gvf_ellipse_par.a" shortname="ell_a" param="GVF_ELLIPSE_A"/>
           <dl_setting MAX="150" MIN="0.0" STEP="5" VAR="gvf_ellipse_par.b" shortname="ell_b" param="GVF_ELLIPSE_B"/>
           <dl_setting MAX="90" MIN="-90" STEP="1" VAR="gvf_ellipse_par.alpha" shortname="ell_alpha" param="GVF_ELLIPSE_ALPHA"/>
         </dl_settings>
-          <dl_settings NAME="Line">
-        <dl_setting MAX="5" MIN="0.0" STEP="0.01" VAR="gvf_line_par.ke" shortname="line_ke" param="GVF_LINE_KE"/>
+        <dl_settings NAME="Line">
+          <dl_setting MAX="50" MIN="0.0" STEP="0.01" VAR="gvf_line_par.ke" shortname="line_ke" param="GVF_LINE_KE"/>
           <dl_setting MAX="5" MIN="0.0" STEP="0.01" VAR="gvf_line_par.kn" shortname="line_kn" param="GVF_LINE_KN"/>
           <dl_setting MAX="180" MIN="-180" STEP="1" VAR="gvf_line_par.heading" shortname="line_heading" param="GVF_LINE_HEADING"/>
           <dl_setting MAX="100" MIN="0" STEP="1" VAR="gvf_segment_par.d1" shortname="d1_seg" param="GVF_SEGMENT_D1"/>
           <dl_setting MAX="100" MIN="0" STEP="1" VAR="gvf_segment_par.d2" shortname="d2_seg" param="GVF_SEGMENT_D2"/>
         </dl_settings>
         <dl_settings NAME="Sine">
-          <dl_setting MAX="5" MIN="0.0" STEP="0.01" VAR="gvf_sin_par.ke" shortname="sin_ke" param="GVF_SIN_KE"/>
+          <dl_setting MAX="50" MIN="0.0" STEP="0.01" VAR="gvf_sin_par.ke" shortname="sin_ke" param="GVF_SIN_KE"/>
           <dl_setting MAX="5" MIN="0.0" STEP="0.01" VAR="gvf_sin_par.kn" shortname="sin_kn" param="GVF_SIN_KN"/>
           <dl_setting MAX="180" MIN="-180" STEP="1" VAR="gvf_sin_par.alpha" shortname="sin_alpha" param="GVF_SIN_ALPHA"/>
           <dl_setting MAX="0.01" MIN="0" STEP="0.0001" VAR="gvf_sin_par.w" shortname="sin_w" param="GVF_SIN_W"/>
@@ -68,10 +68,9 @@ For more details we refer to https://wiki.paparazziuav.org/wiki/Module/guidance_
   
    <header>
      <file name="gvf.h"/>
-     <file name="gvf_low_level_control.h"/>
-     <file name="trajectories/gvf_line.h"/>
-     <file name="trajectories/gvf_sin.h"/>
-     <file name="trajectories/gvf_ellipse.h"/>
+     <file name="nav/nav_line.h"/>
+     <file name="nav/nav_sin.h"/>
+     <file name="nav/nav_ellipse.h"/>
      <file name="nav/nav_survey_polygon_gvf.h"/>
    </header>
 
@@ -79,28 +78,24 @@ For more details we refer to https://wiki.paparazziuav.org/wiki/Module/guidance_
 
    <makefile firmware="fixedwing">
      <file name="gvf.c"/>
-     <file name="gvf_low_level_control.c"/>
-     <file name="trajectories/gvf_line.c"/>
-     <file name="trajectories/gvf_sin.c"/>
-     <file name="trajectories/gvf_ellipse.c"/>
+     <file name="nav/nav_line.c"/>
+     <file name="nav/nav_sin.c"/>
+     <file name="nav/nav_ellipse.c"/>
      <file name="nav/nav_survey_polygon_gvf.c"/>
    </makefile>
    
    <makefile firmware="rotorcraft">
      <file name="gvf.c"/>
-     <file name="gvf_low_level_control.c"/>
-     <file name="trajectories/gvf_line.c"/>
-     <file name="trajectories/gvf_sin.c"/>
-     <file name="trajectories/gvf_ellipse.c"/>
+     <file name="nav/nav_line.c"/>
+     <file name="nav/nav_sin.c"/>
+     <file name="nav/nav_ellipse.c"/>
    </makefile>
 
    <makefile firmware="rover">
      <file name="gvf.c"/>
-     <file name="gvf_low_level_control.c"/>
-     <file name="trajectories/gvf_line.c"/>
-     <file name="trajectories/gvf_sin.c"/>
-     <file name="trajectories/gvf_ellipse.c"/>
-     <!--file name="nav/nav_survey_polygon_gvf.c"/-->
+     <file name="nav/nav_line.c"/>
+     <file name="nav/nav_sin.c"/>
+     <file name="nav/nav_ellipse.c"/>
    </makefile>
 
 </module>
