+++ sw/airborne/modules/guidance/gvf/nav/nav_line.h
@@ -29,7 +29,7 @@
 #ifndef GVF_LINE_H
 #define GVF_LINE_H
 
-#include "modules/guidance/gvf/gvf.h"
+#include "modules/guidance/trajectories/gvf_traj.h"
 
 /** @typedef gvf_li_par
 * @brief Parameters for the GVF line trajectory
@@ -53,9 +53,19 @@ typedef struct {
   float d2;
 } gvf_seg_par;
 
+/** ------------------------------------------------------------------------ **/
+
 extern gvf_li_par gvf_line_par;
 extern gvf_seg_par gvf_segment_par;
 
-extern void gvf_line_info(float *phi, struct gvf_grad *, struct gvf_Hess *);
+// Straight line
+extern bool gvf_line_XY_heading(float x, float y, float heading);
+extern bool gvf_line_XY1_XY2(float x1, float y1, float x2, float y2);
+extern bool gvf_line_wp1_wp2(uint8_t wp1, uint8_t wp2);
+extern bool gvf_segment_loop_XY1_XY2(float x1, float y1, float x2, float y2, float d1, float d2);
+extern bool gvf_segment_loop_wp1_wp2(uint8_t wp1, uint8_t wp2, float d1, float d2);
+extern bool gvf_segment_XY1_XY2(float x1, float y1, float x2, float y2);
+extern bool gvf_segment_wp1_wp2(uint8_t wp1, uint8_t wp2);
+extern bool gvf_line_wp_heading(uint8_t wp, float heading);
 
 #endif // GVF_LINE_H
