+++ sw/airborne/modules/guidance/gvf/nav/nav_survey_polygon_gvf.c
@@ -27,6 +27,8 @@
  *
  */
 
+#include "nav_survey_polygon_gvf.h"
+
 #include "firmwares/fixedwing/nav.h"
 #include "state.h"
 #include "autopilot.h"
