+++ sw/airborne/modules/guidance/gvf_parametric/gvf_parametric.h
@@ -27,33 +27,6 @@
 #ifndef GVF_PARAMETRIC_H
 #define GVF_PARAMETRIC_H
 
-#define GVF_PARAMETRIC_GRAVITY 9.806
-
-/*! Default gain kroll for tuning the "coordinated turn" */
-#ifndef GVF_PARAMETRIC_CONTROL_KROLL
-#define GVF_PARAMETRIC_CONTROL_KROLL 1
-#endif
-
-/*! Default gain kclimb for tuning the climbing setting point */
-#ifndef GVF_PARAMETRIC_CONTROL_KCLIMB
-#define GVF_PARAMETRIC_CONTROL_KCLIMB 1
-#endif
-
-/*! Default scale for the error signals */
-#ifndef GVF_PARAMETRIC_CONTROL_L
-#define GVF_PARAMETRIC_CONTROL_L 0.1
-#endif
-
-/*! Default scale for w  */
-#ifndef GVF_PARAMETRIC_CONTROL_BETA
-#define GVF_PARAMETRIC_CONTROL_BETA 0.01
-#endif
-
-/*! Default gain kpsi for tuning the alignment of the vehicle with the vector field */
-#ifndef GVF_PARAMETRIC_CONTROL_KPSI
-#define GVF_PARAMETRIC_CONTROL_KPSI 1
-#endif
-
 /*! Default GCS trajectory painter */
 #ifndef GVF_OCAML_GCS
 #define GVF_OCAML_GCS true
@@ -63,10 +36,8 @@
 extern "C" {
 #endif
 
-#include "modules/guidance/gvf_parametric/trajectories/gvf_parametric_3d_ellipse.h"
-#include "modules/guidance/gvf_parametric/trajectories/gvf_parametric_3d_lissajous.h"
-#include "modules/guidance/gvf_parametric/trajectories/gvf_parametric_2d_trefoil.h"
-#include "modules/guidance/gvf_parametric/trajectories/gvf_parametric_2d_bezier_splines.h"
+#include "modules/guidance/gvf_common.h"
+#include "gvf_param_traj.h"
 
 /** @typedef gvf_parametric_con
 * @brief Control parameters for the GVF_PARAMETRIC
@@ -80,34 +51,19 @@ typedef struct {
   float w;
   float delta_T;
   int8_t s;
-  float k_roll;
-  float k_climb;
   float k_psi;
   float L;
   float beta;
 } gvf_parametric_con;
 
-extern gvf_parametric_con gvf_parametric_control;
-
-// Parameters for the trajectories
-enum trajectories_parametric {
-  TREFOIL_2D = 0,
-  ELLIPSE_3D = 1,
-  LISSAJOUS_3D = 2,
-  BEZIER_2D = 3,
-  NONE_PARAMETRIC = 255,
-};
-
 typedef struct {
-  enum trajectories_parametric type;
-  float p_parametric[16];
   float phi_errors[3];
-} gvf_parametric_tra;
+  int e_len;
+  uint32_t splines_ctr;
+} gvf_parametric_tel;
 
-extern gvf_parametric_tra gvf_parametric_trajectory;
-
-// Bezier struct
-extern bezier_t gvf_bezier_2D[GVF_PARAMETRIC_2D_BEZIER_N_SEG];
+extern gvf_parametric_con gvf_parametric_control;
+extern gvf_parametric_tel gvf_parametric_telemetry;
 
 // Init function
 extern void gvf_parametric_init(void);
@@ -117,25 +73,6 @@ extern void gvf_parametric_set_direction(int8_t s);
 extern void gvf_parametric_control_2D(float, float, float, float, float, float, float, float);
 extern void gvf_parametric_control_3D(float, float, float, float, float, float, float, float, float,
                                       float, float, float);
-
-// 2D Trefoil
-extern bool gvf_parametric_2D_trefoil_XY(float, float, float, float, float, float, float);
-extern bool gvf_parametric_2D_trefoil_wp(uint8_t, float, float, float, float, float);
-
-// 2D BEZIER
-extern bool gvf_parametric_2D_bezier_wp(uint8_t);
-extern bool gvf_parametric_2D_bezier_XY(void);
-
-// 3D Ellipse
-extern bool gvf_parametric_3D_ellipse_XYZ(float, float, float, float, float, float);
-extern bool gvf_parametric_3D_ellipse_wp(uint8_t, float, float, float, float);
-extern bool gvf_parametric_3D_ellipse_wp_delta(uint8_t, float, float, float, float);
-
-// 3D Lissajous
-extern bool gvf_parametric_3D_lissajous_XYZ(float, float, float, float, float, float, float, float, float, float, float,
-    float, float);
-extern bool gvf_parametric_3D_lissajous_wp_center(uint8_t, float, float, float, float, float, float, float, float,
-    float, float, float);
 
 #ifdef __cplusplus
 }
