+++ sw/airborne/modules/guidance/gvf_parametric/nav/nav_parametric_2d_trefoil.h
@@ -19,47 +19,48 @@
  */
 
 /**
- * @file modules/guidance/gvf_parametric/trajectories/gvf_parametric_3d_ellipse.h
- *
  * Guiding vector field algorithm for 2D and 3D complex trajectories.
  *
- * 3D ellipse (intersection between a cylinder and a tilted plane)
+ * 2D trefoil knot
  */
 
-#ifndef GVF_PARAMETRIC_3D_ELLIPSE_H
-#define GVF_PARAMETRIC_3D_ELLIPSE_H
+#ifndef GVF_PARAMETRIC_2D_TREFOIL_H
+#define GVF_PARAMETRIC_2D_TREFOIL_H
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-/** @typedef gvf_3d_ell_par
-* @brief Parameters for the GVF parametric 3D ellipse
+#include "modules/guidance/gvf_parametric/gvf_param_traj.h"
+
+/** @typedef gvf_2d_tre_par
+* @brief Parameters for the GVF parametric 2D trefoil knot
 * @param kx Gain defining how agressive is the vector field in x coordinate
 * @param ky Gain defining how agressive is the vector field in y coordinate
-* @param kz Gain defining how agressive is the vector field in z coordinate
-* @param r Radius of the cylinder in meters
-* @param zl Altitude of the lowest point of the ellipse
-* @param zh Altitude of the highest point of the ellipse
-* @param alpha Heading of the lowest point zl in rads
+* @param w1 1st frequency
+* @param w2 2nd frequency
+* @param off Off-phase
+* @param r Radius of the "circles"
+* @param alpha Orientation/rotation of the trajectory in the XY plane
 */
 typedef struct {
   float kx;
   float ky;
-  float kz;
+  float w1;
+  float w2;
+  float ratio;
   float r;
-  float zl;
-  float zh;
   float alpha;
-} gvf_par_3d_ell_par;
+} gvf_par_2d_tre_par;
 
-extern gvf_par_3d_ell_par gvf_parametric_3d_ellipse_par;
+extern gvf_par_2d_tre_par gvf_parametric_2d_trefoil_par;
 
-extern void gvf_parametric_3d_ellipse_info(float *f1, float *f2, float *f3, float *f1d, float *f2d, float *f3d,
-    float *f1dd, float *f2dd, float *f3dd);
+// 2D Trefoil
+extern bool gvf_parametric_2D_trefoil_XY(float xo, float yo, float w1, float w2, float ratio, float r, float alpha);
+extern bool gvf_parametric_2D_trefoil_wp(uint8_t wp, float w1, float w2, float ratio, float r, float alpha);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif // GVF_PARAMETRIC_3D_ELLIPSE_H
+#endif // GVF_PARAMETRIC_2D_TREFOIL_H
