+++ sw/airborne/modules/guidance/gvf_parametric/nav/nav_parametric_3d_ellipse.h
@@ -19,46 +19,49 @@
  */
 
 /**
- * @file modules/guidance/gvf_parametric/trajectories/gvf_parametric_2d_trefoil.h
- *
  * Guiding vector field algorithm for 2D and 3D complex trajectories.
  *
- * 2D trefoil knot
+ * 3D ellipse (intersection between a cylinder and a tilted plane)
  */
 
-#ifndef GVF_PARAMETRIC_2D_TREFOIL_H
-#define GVF_PARAMETRIC_2D_TREFOIL_H
+#ifndef GVF_PARAMETRIC_3D_ELLIPSE_H
+#define GVF_PARAMETRIC_3D_ELLIPSE_H
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-/** @typedef gvf_2d_tre_par
-* @brief Parameters for the GVF parametric 2D trefoil knot
+#include "modules/guidance/gvf_parametric/gvf_param_traj.h"
+
+/** @typedef gvf_3d_ell_par
+* @brief Parameters for the GVF parametric 3D ellipse
 * @param kx Gain defining how agressive is the vector field in x coordinate
 * @param ky Gain defining how agressive is the vector field in y coordinate
-* @param w1 1st frequency
-* @param w2 2nd frequency
-* @param off Off-phase
-* @param r Radius of the "circles"
-* @param alpha Orientation/rotation of the trajectory in the XY plane
+* @param kz Gain defining how agressive is the vector field in z coordinate
+* @param r Radius of the cylinder in meters
+* @param zl Altitude of the lowest point of the ellipse
+* @param zh Altitude of the highest point of the ellipse
+* @param alpha Heading of the lowest point zl in rads
 */
 typedef struct {
   float kx;
   float ky;
-  float w1;
-  float w2;
-  float ratio;
+  float kz;
   float r;
+  float zl;
+  float zh;
   float alpha;
-} gvf_par_2d_tre_par;
+} gvf_par_3d_ell_par;
 
-extern gvf_par_2d_tre_par gvf_parametric_2d_trefoil_par;
+extern gvf_par_3d_ell_par gvf_parametric_3d_ellipse_par;
 
-extern void gvf_parametric_2d_trefoil_info(float *f1, float *f2, float *f1d, float *f2d, float *f1dd, float *f2dd);
+// 3D Ellipse
+extern bool gvf_parametric_3D_ellipse_XYZ(float xo, float yo, float r, float zl, float zh, float alpha);
+extern bool gvf_parametric_3D_ellipse_wp(uint8_t wp, float r, float zl, float zh, float alpha);
+extern bool gvf_parametric_3D_ellipse_wp_delta(uint8_t wp, float r, float alt_center, float delta, float alpha);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif // GVF_PARAMETRIC_2D_TREFOIL_H
+#endif // GVF_PARAMETRIC_3D_ELLIPSE_H
